layer: Layer refactorisation.
This is a first step to refactor the haskell layer used in OpenXT:

- Upgrade to GHC 6.12.3, this is a quite old version of GHC, so try at
least to use the latest micro-release of it. There are breaking changes
introduced starting from GHC 7.10 that will require code changes in
Haskell projects used in OpenXT. Upgrading to GHC 7.8 should be and
easier second step. In the interest of testing, this should be done
iteratively. For further upgrades, there is documentation on required
changes at:
https://ghc.haskell.org/trac/ghc/wiki/Migration/7.10

- Patch ghc 6.12.3 to support parallel builds within the OE jethro
branch environment (likely required with later Yocto releases as well).

- Try to comply, or mimic, usual naming and style standards used by
other layers introducing a new language tool chain. I could not find a
documentation for that, yet that makes recipes and bbclasses clearer and
easier to maintain while avoiding redundancies.
Introduce:
    haskell.bbclass: to handle tool chain dependencies and handling for
                     haskell source projects. This is currently using a
                     native compiler instead of a cross tool chain (as
                     OpenXT used to). A cross recipe should be
                     introduced later.
    hackage.bbclass: Factorise common recipe definition to get a package
                     from hackage.haskkell.org.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1213