haskell: ghc local package database
externalsrc.bbclass removes do_fetch, do_unpack and do_patch, so these
cannot be used to set dependencies like haskell.bbclass is doing, or the
class will break under externalsrc.

In the end, the sysroot already is populated with a package database and
bitbake will even handle path substitutions for native targets in the
database (see Staging, stage 2).

There are only minor quirks for target classes that needs to be taken
care of and can be taken care of by prepending the configure step. This
has both advantage of not requiring new steps dependencies and relies on
bitbake managing the sysroot using the sstate to maintain configuration
sanity.

A common example of corruption happens when using externalsrc and
clobbering a version-controlled repository, removing local
configuration files in S.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1317