haskell: remove makeup_wrapper step
The ghc-cc/ghc-ld helper scripts do not need to be generated. Since
haskell.bbclass relies on these for configuration/compilation,
ghc-native can provide a generic script in bindir instead. This is more
coherent as the wrappers belong to the sysroot, not the WORKDIR.

${S}/ghc-{cc,ld} can be externally wiped (externalsrc), which would
remove the wrappers without OE's knowledge. Tracking dynamic files with
the sstate would be asking for trouble as well.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1317