[ghc8] Remove OBE 6.12.3 patches
A majority of the patches have already been implemented upstream.
A few others patched code that doesn't exist anymore. The only
remaining patch (that might not even be needed) is the patch that
fixes a race during install. Repeated testing is needed to verify
that the patch is OBE, and ghc-native takes forever to build, so
we're leaving this patch intact as it doesn't hurt us to do so.

Signed-off-by: Nicholas Tsirakis <niko.tsirakis@gmail.com>