[ghc8] Build GHC with integer-gmp
integer-simple doesn't seem to be widely used anymore; most
libraries will default to gmp unless otherwise specified. In
our case, some new dependencies have a hard depends on gmp,
so we have no choice in the matter.

Note that we don't need to provide libgmp as a DEPENDS as GHC
provides its own in-tree gmp library to use.

Signed-off-by: Nicholas Tsirakis <niko.tsirakis@gmail.com>