[ghc8] Remove usage of deprecated GHC_PACKAGE_PATH env var
GHC 8 now uses a commandline arg to configure this. Since we relied
on this var being exposed in the environment, we now have to pass this
into every binary that cares about it.

Note that we are now recache-ing two separate package databases. It's
unclear why we need to recache both, but failing to recache either one
results in certain packages failing to find the right libraries.

Signed-off-by: Nicholas Tsirakis <niko.tsirakis@gmail.com>