[ghc8] Allow XML tags brought in by external namespaces
Many of our idl interfaces utilize the DBus spec, which introduces
tags that live in the "tp:" namespace. The upgraded dbus library
does not support parsing these tags. This patch ensures we don't
fail early if we come across a tag we can't process.

Note that this does imply that the tp: tags are ignored, but this
change seemingly has no bearing on the generated stubs; the enums
are present despite the new behavior.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>