[ghc8] Add ghc-pkg-wrapper
This is a frustrating amendment to commit a0c71ec422. Some packages
require that we recache both the global package db and the package
db that is local to the repo. For a subset of these packages, the
local db has to also get recached in the middle of the build step
to ensure that any locally-built dependencies are recognized by GHC
when dependencies are resolved later on.

Unfortunately, the compile step is blackboxed in the Setup.hs call,
so we can't recache the local db mid-build without patching GHC. We
also can't pass two -package-db args into the --ghc-pkg-options arg
in the configure step; the last one will always be favored during the
compile step.

To properly get both dbs recached mid-build, we add a ghc-pkg wrapper
that will intercept the ghc-pkg call, check if we're doing a cache
operation, and recache all dbs specified with -package-db if so. If
we aren't doing a recache, just pass the args along to the regular
ghc-pkg.

Note that even though we only pass a single -package-db arg in the
configure step, GHC will automatically supply its own -package-db
arg that points to the local one, so behind the scenes it will have
both arguments to process. Normally, it would only consider the latest
arg supplied (ours), but this commit allows GHC to recache both at
the same time if needed.

Why this is necessary for us is not clear. I imagine it is a side
effect of using OE to orchestrate the build. Either way, it's not
an ideal solution, but it's the only one I found that works.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>